update text field: separate formDemo | add error message next to input
